requirements.txt
# RAPIDS GPU Data Processing
cudf-cu12>=24.02
cuml-cu12>=24.02
dask-cudf-cu12>=24.02
dask[distributed]>=2024.1.0

# XGBoost with GPU
xgboost>=2.0.0

# Data formats
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# API & Serving
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Experiment Tracking
mlflow>=2.10.0

# Hyperparameter Tuning
optuna>=3.5.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Notebooks
jupyterlab>=4.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
